Using BesuLocalNode instead of docker
