+++ acceptance-tests/build.gradle
@@ -81,6 +81,11 @@ task acceptanceTest(type: Test) {
   systemProperty 'java.security.properties', "${buildDir}/resources/test/acceptanceTesting.security"
   systemProperty 'besuInstallDir', "${buildDir}/besu"
 
+  // CircleCI - Hashicorp tests are run in a docker_machine environment separately
+  if (project.hasProperty("excludeHashicorp")) {
+    exclude 'tech/pegasys/ethsigner/tests/signing/hashicorp/**'
+  }
+
   useJUnitPlatform()
   // toggle to show standard out and standard error of the test JVM(s) on the console
   testLogging.showStandardStreams = false
