+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/dsl/node/Node.java
@@ -25,6 +25,8 @@ public interface Node {
 
   void awaitStartupCompletion();
 
+  String hostName();
+
   NodePorts ports();
 
   Accounts accounts();
