+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/multikeysigner/transactionsigning/MultiKeyTransactionSigningAcceptanceTestBase.java
@@ -17,11 +17,9 @@ import static tech.pegasys.ethsigner.tests.dsl.Gas.GAS_PRICE;
 import static tech.pegasys.ethsigner.tests.dsl.Gas.INTRINSIC_GAS;
 
 import tech.pegasys.ethsigner.tests.dsl.Account;
-import tech.pegasys.ethsigner.tests.dsl.DockerClientFactory;
-import tech.pegasys.ethsigner.tests.dsl.node.BesuDockerNode;
 import tech.pegasys.ethsigner.tests.dsl.node.Node;
-import tech.pegasys.ethsigner.tests.dsl.node.NodeConfiguration;
-import tech.pegasys.ethsigner.tests.dsl.node.NodeConfigurationBuilder;
+import tech.pegasys.ethsigner.tests.dsl.node.besu.BesuLocalNodeFactory;
+import tech.pegasys.ethsigner.tests.dsl.node.besu.BesuNodeConfigBuilder;
 import tech.pegasys.ethsigner.tests.dsl.signer.Signer;
 import tech.pegasys.ethsigner.tests.dsl.signer.SignerConfiguration;
 import tech.pegasys.ethsigner.tests.dsl.signer.SignerConfigurationBuilder;
@@ -30,7 +28,6 @@ import tech.pegasys.ethsigner.tests.multikeysigner.MultiKeyAcceptanceTestBase;
 import java.math.BigInteger;
 import java.nio.file.Path;
 
-import com.github.dockerjava.api.DockerClient;
 import org.junit.jupiter.api.AfterAll;
 import org.web3j.protocol.core.methods.request.Transaction;
 import org.web3j.utils.Convert;
@@ -50,17 +47,17 @@ public class MultiKeyTransactionSigningAcceptanceTestBase extends MultiKeyAccept
     Runtime.getRuntime()
         .addShutdownHook(new Thread(MultiKeyTransactionSigningAcceptanceTestBase::tearDownBase));
 
-    final DockerClient docker = new DockerClientFactory().create();
-    final NodeConfiguration nodeConfig = new NodeConfigurationBuilder().build();
+    // final DockerClient docker = new DockerClientFactory().create();
+    // final NodeConfiguration nodeConfig = new NodeConfigurationBuilder().build();
 
-    ethNode = new BesuDockerNode(docker, nodeConfig);
+    ethNode = new BesuLocalNodeFactory().create(BesuNodeConfigBuilder.aBesuNodeConfig().build());
     ethNode.start();
     ethNode.awaitStartupCompletion();
 
     final SignerConfiguration signerConfig =
         new SignerConfigurationBuilder().withMultiKeySignerDirectory(tomlDirectory).build();
 
-    ethSigner = new Signer(signerConfig, nodeConfig.getHostname(), ethNode.ports());
+    ethSigner = new Signer(signerConfig, ethNode.hostName(), ethNode.ports());
     ethSigner.start();
     ethSigner.awaitStartupCompletion();
   }
